Move checkout-calculate-taxes to checkout domain (#18967)
