+++ saleor/plugins/webhook/tests/test_tax_webhook.py
@@ -1,40 +1,8 @@
-import json
-from unittest import mock
-from unittest.mock import ANY, sentinel
+from unittest.mock import sentinel
 
 import pytest
-from freezegun import freeze_time
 
-from ....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
-from ....core import EventDeliveryStatus
-from ....core.models import EventDelivery
-from ....core.taxes import TaxDataError, TaxType
-from ....graphql.webhook.utils import get_subscription_query_hash
-from ....tax.webhooks.parser import parse_tax_data
-from ....webhook.event_types import WebhookEventSyncType
-from ...manager import get_plugins_manager
-
-
-@mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
-def test_get_taxes_for_checkout_no_permission(
-    mock_request,
-    webhook_plugin,
-    checkout,
-):
-    # given
-    plugin = webhook_plugin()
-    lines, _ = fetch_checkout_lines(checkout)
-    checkout_info = fetch_checkout_info(
-        checkout, lines, get_plugins_manager(allow_replica=False)
-    )
-    app_identifier = None
-
-    # when
-    tax_data = plugin.get_taxes_for_checkout(checkout_info, lines, app_identifier, None)
-
-    # then
-    assert mock_request.call_count == 0
-    assert tax_data is None
+from ....core.taxes import TaxType
 
 
 @pytest.fixture
@@ -96,270 +64,3 @@ def test_get_tax_code_from_object_meta_default_code(
         code="UNMAPPED",
         description="Unmapped Product/Product Type",
     )
-
-
-@freeze_time()
-@mock.patch("saleor.checkout.calculations.fetch_checkout_data")
-@mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
-@mock.patch(
-    "saleor.webhook.transport.synchronous.transport.generate_payload_from_subscription"
-)
-def test_get_taxes_for_checkout_with_sync_subscription(
-    mock_generate_payload,
-    mock_request,
-    mock_fetch,
-    webhook_plugin,
-    tax_data_response,
-    checkout,
-    tax_app,
-):
-    # given
-    subscription_query = "subscription{event{... on CalculateTaxes{taxBase{currency}}}}"
-    expected_payload = {"taxBase": {"currency": "USD"}}
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    mock_request.return_value = tax_data_response
-    mock_generate_payload.return_value = expected_payload
-    plugin = webhook_plugin()
-    webhook = tax_app.webhooks.get(name="tax-webhook-1")
-    webhook.subscription_query = subscription_query
-    webhook.save(update_fields=["subscription_query"])
-    app_identifier = None
-
-    # when
-    tax_data = plugin.get_taxes_for_checkout(checkout_info, [], app_identifier, None)
-
-    # then
-    mock_generate_payload.assert_called_once_with(
-        event_type=WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-        subscribable_object=checkout,
-        subscription_query=subscription_query,
-        request=ANY,  # SaleorContext,
-    )
-    mock_request.assert_called_once()
-    assert not EventDelivery.objects.exists()
-
-    delivery = mock_request.mock_calls[0].args[0]
-    assert delivery.payload.get_payload() == json.dumps(expected_payload)
-    assert delivery.status == EventDeliveryStatus.PENDING
-    assert delivery.event_type == WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES
-    assert delivery.webhook == webhook
-    mock_fetch.assert_not_called()
-    assert tax_data == parse_tax_data(
-        WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-        tax_data_response,
-        checkout.lines.count(),
-    )
-
-
-@freeze_time()
-@mock.patch("saleor.checkout.calculations.fetch_checkout_data")
-@mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
-@mock.patch(
-    "saleor.webhook.transport.synchronous.transport.generate_payload_from_subscription"
-)
-def test_get_taxes_for_checkout_with_sync_subscription_with_pregenerated_payload(
-    mock_generate_payload,
-    mock_request,
-    mock_fetch,
-    webhook_plugin,
-    tax_data_response,
-    checkout,
-    tax_app,
-):
-    # given
-    subscription_query = "subscription{event{... on CalculateTaxes{taxBase{currency}}}}"
-    subscription_query_hash = get_subscription_query_hash(subscription_query)
-    expected_payload = {"taxBase": {"currency": "USD"}}
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    mock_request.return_value = tax_data_response
-    plugin = webhook_plugin()
-    webhook = tax_app.webhooks.get(name="tax-webhook-1")
-    webhook.subscription_query = subscription_query
-    webhook.save(update_fields=["subscription_query"])
-    app_identifier = None
-    pregenerated_subscription_payloads = {
-        webhook.app.id: {subscription_query_hash: expected_payload}
-    }
-
-    # when
-    tax_data = plugin.get_taxes_for_checkout(
-        checkout_info,
-        [],
-        app_identifier,
-        None,
-        pregenerated_subscription_payloads=pregenerated_subscription_payloads,
-    )
-
-    # then
-    mock_generate_payload.assert_not_called()
-    mock_request.assert_called_once()
-    assert not EventDelivery.objects.exists()
-
-    delivery = mock_request.mock_calls[0].args[0]
-    assert delivery.payload.get_payload() == json.dumps(expected_payload)
-    assert delivery.status == EventDeliveryStatus.PENDING
-    assert delivery.event_type == WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES
-    assert delivery.webhook == webhook
-    mock_fetch.assert_not_called()
-    assert tax_data == parse_tax_data(
-        WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-        tax_data_response,
-        checkout.lines.count(),
-    )
-
-
-def test_get_taxes_for_checkout_with_app_identifier_app_missing(
-    checkout_info, webhook_plugin
-):
-    # given
-    app_identifier = "missing_app"
-    plugin = webhook_plugin()
-
-    # when & then
-    with pytest.raises(TaxDataError, match="Configured tax app doesn't exist."):
-        plugin.get_taxes_for_checkout(
-            checkout_info, checkout_info.lines, app_identifier, None
-        )
-
-
-def test_get_taxes_for_checkout_with_app_identifier_webhook_is_missing(
-    checkout_info, webhook_plugin, app
-):
-    # given
-    plugin = webhook_plugin()
-
-    # when & then
-    with pytest.raises(
-        TaxDataError,
-        match="Configured tax app's webhook for taxes calculation doesn't exists.",
-    ):
-        plugin.get_taxes_for_checkout(
-            checkout_info, checkout_info.lines, app.identifier, None
-        )
-
-
-@mock.patch("saleor.checkout.calculations.fetch_checkout_data")
-@mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
-@mock.patch(
-    "saleor.webhook.transport.synchronous.transport.generate_payload_from_subscription"
-)
-def test_get_taxes_for_checkout_with_app_identifier_invalid_response(
-    mock_generate_payload,
-    mock_request,
-    mock_fetch,
-    webhook_plugin,
-    tax_data_response_factory,
-    checkout,
-    tax_app,
-):
-    # given
-    subscription_query = "subscription{event{... on CalculateTaxes{taxBase{currency}}}}"
-    expected_payload = {"taxBase": {"currency": "USD"}}
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    # mock returning invalid data - tax_rate with negative value
-    mock_request.return_value = tax_data_response_factory(shipping_tax_rate=-10)
-    mock_generate_payload.return_value = expected_payload
-    plugin = webhook_plugin()
-    webhook = tax_app.webhooks.get(name="tax-webhook-1")
-    webhook.subscription_query = subscription_query
-    webhook.save(update_fields=["subscription_query"])
-    app_identifier = tax_app.identifier
-
-    # when
-    with pytest.raises(TaxDataError):
-        plugin.get_taxes_for_checkout(checkout_info, [], app_identifier, None)
-
-
-@freeze_time()
-@mock.patch("saleor.checkout.calculations.fetch_checkout_data")
-@mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
-@mock.patch(
-    "saleor.webhook.transport.synchronous.transport.generate_payload_from_subscription"
-)
-def test_get_taxes_for_checkout_with_app_identifier(
-    mock_generate_payload,
-    mock_request,
-    mock_fetch,
-    webhook_plugin,
-    tax_data_response,
-    checkout,
-    tax_app,
-):
-    # given
-    subscription_query = "subscription{event{... on CalculateTaxes{taxBase{currency}}}}"
-    expected_payload = {"taxBase": {"currency": "USD"}}
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    mock_request.return_value = tax_data_response
-    mock_generate_payload.return_value = expected_payload
-    plugin = webhook_plugin()
-    webhook = tax_app.webhooks.get(name="tax-webhook-1")
-    webhook.subscription_query = subscription_query
-    webhook.save(update_fields=["subscription_query"])
-    app_identifier = tax_app.identifier
-
-    # when
-    tax_data = plugin.get_taxes_for_checkout(checkout_info, [], app_identifier, None)
-
-    # then
-    mock_generate_payload.assert_called_once_with(
-        event_type=WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-        subscribable_object=checkout,
-        subscription_query=subscription_query,
-        request=ANY,  # SaleorContext,
-    )
-    mock_request.assert_called_once()
-    assert not EventDelivery.objects.exists()
-
-    delivery = mock_request.mock_calls[0].args[0]
-    assert delivery.payload.get_payload() == json.dumps(expected_payload)
-    assert delivery.status == EventDeliveryStatus.PENDING
-    assert delivery.event_type == WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES
-    assert delivery.webhook == webhook
-    mock_fetch.assert_not_called()
-    assert tax_data == parse_tax_data(
-        WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-        tax_data_response,
-        checkout.lines.count(),
-    )
-
-
-@freeze_time()
-@mock.patch("saleor.checkout.calculations.fetch_checkout_data")
-@mock.patch("saleor.webhook.transport.synchronous.transport.send_webhook_request_sync")
-@mock.patch(
-    "saleor.webhook.transport.synchronous.transport.generate_payload_from_subscription"
-)
-def test_get_taxes_for_checkout_with_app_identifier_empty_response(
-    mock_generate_payload,
-    mock_request,
-    mock_fetch,
-    webhook_plugin,
-    tax_data_response,
-    checkout,
-    tax_app,
-):
-    # given
-    subscription_query = "subscription{event{... on CalculateTaxes{taxBase{currency}}}}"
-    expected_payload = {"taxBase": {"currency": "USD"}}
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    mock_request.return_value = None
-    mock_generate_payload.return_value = expected_payload
-    plugin = webhook_plugin()
-    webhook = tax_app.webhooks.get(name="tax-webhook-1")
-    webhook.subscription_query = subscription_query
-    webhook.save(update_fields=["subscription_query"])
-    app_identifier = tax_app.identifier
-
-    # when & then
-    with pytest.raises(TaxDataError):
-        plugin.get_taxes_for_checkout(checkout_info, [], app_identifier, None)
